# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for the B2B medical supplies marketplace
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products/{id}/variants:
    get:
      operationId: listProductVariants
      tags: [variants]
      summary: List variants for a master product (e.g. sizes)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Variants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductVariant"
  /admin/products/{id}/variants:
    post:
      operationId: createProductVariant
      tags: [variants]
      summary: Add a variant to a master product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductVariantInput"
      responses:
        "201":
          description: Variant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductVariant"
  /admin/variants/{id}:
    delete:
      operationId: deleteProductVariant
      tags: [variants]
      summary: Remove a product variant (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /vendor/warehouses:
    get:
      operationId: listMyWarehouses
      tags: [warehouses]
      summary: List the current vendor's warehouses
      responses:
        "200":
          description: Warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warehouse"
    post:
      operationId: createWarehouse
      tags: [warehouses]
      summary: Create a warehouse for the current vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseInput"
      responses:
        "201":
          description: Warehouse created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
  /vendor/warehouses/{id}:
    patch:
      operationId: updateWarehouse
      tags: [warehouses]
      summary: Update a warehouse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
    delete:
      operationId: deleteWarehouse
      tags: [warehouses]
      summary: Delete a warehouse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/companies/{id}/scorecard:
    get:
      operationId: getVendorScorecard
      tags: [scorecards]
      summary: Performance scorecard for a given vendor (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scorecard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorScorecard"
...
components:
...
  schemas:
...
    ProductVariant:
      type: object
      required: [id, productId, name, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        name:
          type: string
        sku:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
    ProductVariantInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        sku:
          type: string
    Warehouse:
      type: object
      required: [id, vendorCompanyId, name, createdAt]
      properties:
        id:
          type: integer
        vendorCompanyId:
          type: integer
        name:
          type: string
        address:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
    WarehouseInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        address:
          type: string
...
    VendorScorecard:
      type: object
      required:
        [
          vendorCompanyId,
          fulfillmentRate,
          cancellationRate,
          avgDeliveryDays,
          avgRating,
          totalOrders,
          totalReviews,
        ]
      properties:
        vendorCompanyId:
          type: integer
        fulfillmentRate:
          type: number
        cancellationRate:
          type: number
        avgDeliveryDays:
          type: number
        avgRating:
          type: number
        totalOrders:
          type: integer
        totalReviews:
          type: integer
...